requirements.txt
fastapi
uvicorn[standard]
markdown
aiosmtplib
jinja2
python-dotenv
